+++ src/core/geometry/primitive.h
@@ -1,12 +1,13 @@
 #pragma once
 
 #include <vector>
+#include <cstdint>
 
 namespace butter {
 
 struct Primitive{
 
-    std::vector<size_t> vertices; // list of vertex indices
+    std::vector<uint32_t> vertices; // list of vertex indices
 
 };
